+++ package.json
@@ -31,6 +31,7 @@
     "react-calendar": "^4.7.0",
     "react-cookie": "6.1.1",
     "react-dom": "18.2.0",
+    "react-hook-form": "^7.52.2",
     "styled-components": "6.1.0",
     "swiper": "11.0.5"
   },
