+++ src/app/signout/page.tsx
@@ -2,7 +2,7 @@
 import useFunnel from '@/hooks/useFunnel';
 import { SignOutInfoProvider } from '@/app/signout/signoutContext';
 
-import { SignOutInfo } from '@/app/signout/(components)/signout-info';
+import { SignOutTypeSelect } from '@/app/signout/(components)/signout-type-select';
 import { SignOutHeader } from '@/app/signout/(components)/Header';
 
 const signOutSteps = [
@@ -22,6 +22,7 @@ export default function SignOut() {
       <SignOutInfoProvider>
         <SignOutHeader onClick={() => prevStep()} />
         <SignoutFunnel steps={signOutSteps} step="signout_info">
+          //TODO - 회원 탈퇴 문구와 함께 UI 수정 필요
           <SignoutFunnel.Step name={'signout_info'}>
             <div>회원 탈퇴 공지</div>
             <button onClick={() => setSignoutStep('signout_type_select')}>
@@ -29,7 +30,9 @@ export default function SignOut() {
             </button>
           </SignoutFunnel.Step>
           <SignoutFunnel.Step name={'signout_type_select'}>
-            <SignOutInfo onClick={() => setSignoutStep('signout_reason')} />
+            <SignOutTypeSelect
+              onClick={() => setSignoutStep('signout_reason')}
+            />
           </SignoutFunnel.Step>
           <SignoutFunnel.Step name={'signout_reason'}>
             <div>회원 탈퇴 이유</div>
